8217500: (sc) Move SocketChannelImpl's remaining native methods to Net Reviewed-by: bpb
