+++ test/jdk/java/nio/channels/SocketChannel/SendUrgentData.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2015, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -133,9 +133,7 @@ public class SendUrgentData {
                             throw new RuntimeException("Unexpected message", ex);
                         }
                     } else if (osName.contains("windows")) {
-                        if (!(ex instanceof SocketException)) {
-                            throw new RuntimeException("Unexpected exception", ex);
-                        } else if (!ex.getMessage().contains("Resource temporarily unavailable")) {
+                        if (!ex.getMessage().equals("Socket buffer full")) {
                             throw new RuntimeException("Unexpected message", ex);
                         }
                     } else {
